Flyttet metoder fra taskworker til TaskStepService for å hente ut eks maxAntallFeilMap
