+++ prosessering-core/src/main/kotlin/no/nav/familie/prosessering/internal/TaskWorker.kt
@@ -1,46 +1,15 @@
 package no.nav.familie.prosessering.internal
 
-import io.micrometer.core.instrument.Counter
-import io.micrometer.core.instrument.Metrics
-import no.nav.familie.prosessering.AsyncTaskStep
 import no.nav.familie.prosessering.TaskFeil
-import no.nav.familie.prosessering.TaskStepBeskrivelse
 import no.nav.familie.prosessering.domene.ITask
 import no.nav.familie.prosessering.domene.Status
 import org.slf4j.LoggerFactory
-import org.springframework.aop.framework.AopProxyUtils
-import org.springframework.core.annotation.AnnotationUtils
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Propagation
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class TaskWorker(private val taskRepository: TaskService, taskStepTyper: List<AsyncTaskStep>) {
-
-    private val taskStepMap: Map<String, AsyncTaskStep>
-
-    private val maxAntallFeilMap: Map<String, Int>
-    private val triggerTidVedFeilMap: Map<String, Long>
-    private val feiltellereForTaskSteps: Map<String, Counter>
-
-    init {
-        val tasksTilTaskStepBeskrivelse: Map<AsyncTaskStep, TaskStepBeskrivelse> = taskStepTyper.associateWith { task ->
-            val aClass = AopProxyUtils.ultimateTargetClass(task)
-            val annotation = AnnotationUtils.findAnnotation(aClass, TaskStepBeskrivelse::class.java)
-            requireNotNull(annotation) { "annotasjon mangler" }
-            annotation
-        }
-        taskStepMap = tasksTilTaskStepBeskrivelse.entries.associate { it.value.taskStepType to it.key }
-        maxAntallFeilMap = tasksTilTaskStepBeskrivelse.values.associate { it.taskStepType to it.maxAntallFeil }
-        triggerTidVedFeilMap = tasksTilTaskStepBeskrivelse.values.associate { it.taskStepType to it.triggerTidVedFeilISekunder }
-        feiltellereForTaskSteps = tasksTilTaskStepBeskrivelse.values.associate {
-            it.taskStepType to Metrics.counter("mottak.feilede.tasks",
-                                               "status",
-                                               it.taskStepType,
-                                               "beskrivelse",
-                                               it.beskrivelse)
-        }
-    }
+class TaskWorker(private val taskRepository: TaskService, private val taskStepService: TaskStepService) {
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     fun doActualWork(taskId: Long) {
@@ -54,7 +23,7 @@ class TaskWorker(private val taskRepository: TaskService, taskStepTyper: List<As
         task = task.behandler()
         task = taskRepository.save(task)
         // finn tasktype
-        val taskStep = finnTaskStep(task.type)
+        val taskStep = taskStepService.finnTaskStep(task.type)
 
         // execute
         taskStep.preCondition(task)
@@ -71,37 +40,19 @@ class TaskWorker(private val taskRepository: TaskService, taskStepTyper: List<As
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     fun doFeilhåndtering(taskId: Long, e: Exception) {
         var task = taskRepository.findById(taskId)
-        val maxAntallFeil = finnMaxAntallFeil(task.type)
+        val maxAntallFeil = taskStepService.finnMaxAntallFeil(task.type)
         secureLog.trace("Behandler task='{}'", task)
 
         task = task.feilet(TaskFeil(task, e), maxAntallFeil)
         // lager metrikker på tasks som har feilet max antall ganger.
         if (task.status == Status.FEILET) {
-            finnFeilteller(task.type).increment()
+            taskStepService.finnFeilteller(task.type).increment()
             log.error("Task ${task.id} av type ${task.type} har feilet. " +
                       "Sjekk familie-prosessering for detaljer")
         }
-        task = task.medTriggerTid(task.triggerTid.plusSeconds(finnTriggerTidVedFeil(task.type)))
+        task = task.medTriggerTid(task.triggerTid.plusSeconds(taskStepService.finnTriggerTidVedFeil(task.type)))
         taskRepository.save(task)
         secureLog.info("Feilhåndtering lagret ok {}", task)
-
-    }
-
-
-    private fun finnTriggerTidVedFeil(taskType: String): Long {
-        return triggerTidVedFeilMap[taskType] ?: 0
-    }
-
-    private fun finnFeilteller(taskType: String): Counter {
-        return feiltellereForTaskSteps[taskType] ?: error("Ukjent tasktype $taskType")
-    }
-
-    private fun finnMaxAntallFeil(taskType: String): Int {
-        return maxAntallFeilMap[taskType] ?: error("Ukjent tasktype $taskType")
-    }
-
-    private fun finnTaskStep(taskType: String): AsyncTaskStep {
-        return taskStepMap[taskType] ?: error("Ukjent tasktype $taskType")
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
